Extracted editor window toolbars to separate controller file.
