+++ src/main/java/com/codecool/am_i_tea/AmITea.java
@@ -1,5 +1,6 @@
 package com.codecool.am_i_tea;
 
+import com.codecool.am_i_tea.controller.EditorController;
 import com.codecool.am_i_tea.controller.EditorMenuController;
 import com.codecool.am_i_tea.dao.ProjectDAO;
 import com.codecool.am_i_tea.dao.TextFileDAO;
@@ -7,21 +8,12 @@ import com.codecool.am_i_tea.service.LoggerService;
 import com.codecool.am_i_tea.service.ProjectService;
 import com.codecool.am_i_tea.service.PropertyUtil;
 import com.codecool.am_i_tea.service.TextFileService;
-import com.codecool.paintFx.model.ShapeList;
 import com.codecool.paintFx.service.PaintService;
 import javafx.application.Application;
-import javafx.application.Platform;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.fxml.FXMLLoader;
-import javafx.scene.Node;
 import javafx.scene.Scene;
 import javafx.scene.canvas.Canvas;
 import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.control.*;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.input.MouseButton;
 import javafx.scene.layout.*;
 import javafx.scene.web.HTMLEditor;
 import javafx.scene.web.WebView;
@@ -29,9 +21,7 @@ import javafx.stage.Stage;
 
 import netscape.javascript.JSObject;
 
-import javax.swing.*;
 import java.io.*;
-import java.util.List;
 import java.util.Properties;
 
 
@@ -49,6 +39,7 @@ public class AmITea extends Application {
     private GraphicsContext graphicsContext;
     private LoggerService logger;
     private EditorMenuController editorMenuController;
+    private EditorController editorController;
 
     public static void main(String[] args) {
         launch(args);
@@ -77,9 +68,14 @@ public class AmITea extends Application {
         //todo: this is not in the right order
         HTMLEditor editor = new HTMLEditor();
         editor.setVisible(false);
+        WebView webView = (WebView) editor.lookup("WebView");
 
         editorMenuController = new EditorMenuController(logger, propertyUtil, projectService,
                 textFileService, projectDAO, fileDAO, primaryStage, editor, graphicsContext);
+        editorMenuController.initializeMenuBar();
+
+        editorController = new EditorController(editor, wrapper, webView);
+        editorController.addButtonsToEditorToolbar();
 
         logger.log("AmITea application started!");
 
@@ -87,8 +83,6 @@ public class AmITea extends Application {
 
 
 
-
-        WebView webView = (WebView) editor.lookup("WebView");
         JavaApplication javaApp = new JavaApplication(fileDAO, textFileService, projectDAO, editor);
 
         webView.getEngine().setJavaScriptEnabled(true);
@@ -100,52 +94,7 @@ public class AmITea extends Application {
         primaryStage.setTitle("Am-I-Tea text editor");
 
 
-        editorMenuController.initializeMenuBar();
-        MenuBar menuBar = editorMenuController.getMenuBar();
 
-        Node node = editor.lookup(".top-toolbar");
-        if (node instanceof ToolBar) {
-            ToolBar bar = (ToolBar) node;
-
-            Button drawButton = new Button("Drawing mode");
-            drawButton.setTooltip(new Tooltip("Draw"));
-
-            Button linkButton = new Button("Hyperlink");
-            ImageView linkImageView = new ImageView(new Image(getClass().getResourceAsStream("/images/hyperlink.png")));
-            linkButton.setMinSize(26.0, 22.0);
-            linkButton.setMaxSize(26.0, 22.0);
-            linkImageView.setFitHeight(16);
-            linkImageView.setFitWidth(16);
-            linkButton.setContentDisplay(ContentDisplay.GRAPHIC_ONLY);
-            linkButton.setGraphic(linkImageView);
-            linkButton.setTooltip(new Tooltip("Link another file"));
-
-            linkButton.setOnAction(actionEvent -> {
-                String targetFileName = JOptionPane.showInputDialog("Enter file name");
-
-                String selected = (String) webView.getEngine().executeScript("window.getSelection().toString();");
-                selected = formatSelection(selected);
-                String hyperlinkHtml = "<span style=\"color:blue; text-decoration:underline; \" onClick=\"" +
-                        "window.app.openLinkedFile(\\'" + targetFileName + "\\')\"" + ">" + selected + "</span>";
-                webView.getEngine().executeScript(getInsertHtmlAtCursorJS(hyperlinkHtml));
-            });
-
-            drawButton.setOnAction(actionEvent -> {
-                wrapper.getChildren().get(1).setMouseTransparent(false);
-                Node topToolBar = editor.lookup(".top-toolbar");
-                Node bottomToolBar = editor.lookup(".bottom-toolbar");
-                topToolBar.setVisible(false);
-                bottomToolBar.setVisible(false);
-                showDrawSceneToolBars(true);
-            });
-
-            Separator separator = new Separator();
-
-            bar.getItems().add(drawButton);
-            bar.getItems().add(separator);
-            bar.getItems().add(linkButton);
-
-        }
         try {
             drawScene = new Scene(FXMLLoader.load(getClass().getClassLoader().getResource("paint.fxml")));
         } catch (IOException e) {
@@ -165,9 +114,9 @@ public class AmITea extends Application {
 
         VBox wrapperVbox = new VBox();
         Scene wrapperScene = new Scene(wrapperVbox);
-        ((VBox) wrapperScene.getRoot()).getChildren().addAll(menuBar, wrapper);
+        ((VBox) wrapperScene.getRoot()).getChildren().addAll(editorMenuController.getMenuBar(), wrapper);
 
-        showDrawSceneToolBars(false);
+        editorController.showDrawSceneToolBars(false);
 
         graphicsContext = ((Canvas) drawScene.getRoot().getChildrenUnmodifiable().get(1)).getGraphicsContext2D();
         textFileService.setGraphicsContext(graphicsContext);
@@ -176,37 +125,5 @@ public class AmITea extends Application {
         primaryStage.show();
 
 
-    }
-
-    private String formatSelection(String selected) {
-        selected = selected.replaceAll("\\n\\n", "<br>");
-        return selected.replaceAll("\\\\", "\\\\\\\\");
-    }
-
-    private void showDrawSceneToolBars(Boolean show) {
-        Node myDrawNode = wrapper.getChildren().get(1);
-        BorderPane myDrawScene = (BorderPane) myDrawNode;
-        VBox myVbox = (VBox) myDrawScene.getChildren().get(0);
-        Node topToolBar = myVbox.getChildren().get(0);
-        Node bottomToolBar = myVbox.getChildren().get(1);
-        topToolBar.setVisible(show);
-        bottomToolBar.setVisible(show);
-
-    }
-
-    private String getInsertHtmlAtCursorJS(String html) {
-        return "insertHtmlAtCursor('" + html + "');"
-                + "function insertHtmlAtCursor(html) {\n"
-                + " var range, node;\n"
-                + " if (window.getSelection && window.getSelection().getRangeAt) {\n"
-                + " window.getSelection().deleteFromDocument();\n"
-                + " range = window.getSelection().getRangeAt(0);\n"
-                + " node = range.createContextualFragment(html);\n"
-                + " range.insertNode(node);\n"
-                + " } else if (document.selection && document.selection.createRange) {\n"
-                + " document.selection.createRange().pasteHTML(html);\n"
-                + " document.selection.clear();"
-                + " }\n"
-                + "}";
     }
 }
